Align the customer Jobs-tab New Job modal with the main Jobs form (#54)
Closes #54.

The customer edit screen's Jobs tab used a stripped-down New Job modal
(property, title, status, priority, date, notes) that diverged from the real
Jobs create form. It now reuses JobResource::form() verbatim, so the modal
carries the same fields and tabs — including the Services tab with TBD pricing
and the recurrence options. The customer is pinned to the tab's owner.

The one-time/recurring + service-line creation logic that lived on the CreateJob
page is extracted into JobFromFormCreator, shared by the page and the modal so
both behave identically.
